refactor(gr-utils): add all geographic regions
